% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xsdata.R
\name{get_xsdata_bin}
\alias{get_xsdata_bin}
\title{Locate xsdata utility.}
\usage{
get_xsdata_bin()
}
\description{
This auxiliary function will check whether xsdata is on the path, available
in the directory \verb{/opt/homebrew/bin} or at the environment variable \code{XSDATA_BIN}.
}
\details{
If xsdata is not available, the recommended approach is to call \verb{pip install xsdata[cli,lxml,soap]}. Optionally, try \verb{pip3 install xsdata[cli,lxml,soap]}.
See \href{https://xsdata.readthedocs.io/en/latest/installation.html}{xsdata installation instructions}
}
