% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruml.R
\name{class_get_subclasses}
\alias{class_get_subclasses}
\alias{class_get_subclasses_vec}
\alias{pkg_get_all_methods}
\title{Auxiliary functions.}
\usage{
class_get_subclasses(x)

class_get_subclasses_vec(x)

pkg_get_all_methods(pkg)
}
\arguments{
\item{x}{A class (length-one \code{character} vector).}

\item{pkg}{Unquoted package name.}
}
\description{
Auxiliary functions.
}
\details{
\code{class_get_subclasses} will retur a two-column
\code{data.frame} (columns "super" and "sub") with the relations between
super- and subclasses.

\code{class_get_subclasses_vec} will return a \code{character} vector with
all subclasses that inherit from \code{x}.
}
\examples{
library(polmineR)
class_get_subclasses("count")
class_get_subclasses_vec
class_get_subclasses_vec("textstat")
pkg_get_all_methods(polmineR)
}
